feat: prettier, re-factor, folder-structure
