+++ packages/react-tagify/.prettierrc
